Aalen model's description
We know Cox proportional hazard model is a very popular model in applications, especially in medical research.
Actually, additive hazard model is  an important alternative to the proportional hazards model, which assumes that the covariates act in an additive manner on an unknown baseline hazard rate. That is:
h(t|z)=h0(t)+b(t)'z(t)
Related models have been investigated by Aalen (1980, 1989), Huffer and McKeague (1991), Lin and Ying (1994), Lin and Ying
(1995), and McKeague and Sasieni (1994), among others. In particular, Lin and Ying (1994)
proposed a score-type function to obtain an estimator of 0 in an analytic form.